major concept refactor, they all work now
concepts are working again, and actually have tests in detect.cpp, there
are new macros for defining detectors, requirements (helper for
detectors), concepts, and all-in-one doing a concept and associated
traits.  It's possible there are breaking changes in here if anyone
external is relying on the concepts, but since many of them were broken
I'm hoping this will be a net benefit.