use integer_pack, make_integer_seq and type_pack_e
We were already using __make_integer_seq, incorporate the __integer_pack
code from the concepts refactor branch and add cases for intel, nvcc and
pgi to be able to use one or more of these intrinsics as well.  It turns
out EDG supports __make_integer_seq and __integer_pack, but only when configured some
certain way.  NVCC supports both, with the clang builtin supported only
when clang is selected as a host compiler and __integer_pack supported
only in version 10+ and only with a g++ >= 8 selected as host compiler.
Both of PGI and Intel seem to do something similar, supporting only
__integer_pack, and only when their targeted compiler and libstdc++ is
8+.  I have yet to try XL, but that's next.